Migration to cmgui r7020 finished

+++ src/CAPModelLVPS4X4.h
@@ -110,7 +110,7 @@ public:
 		numberOfModelFrames_ = numberOfFrames;
 	}
 	
-	float ComputeVolume(SurfaceType surface, float time) const;
+	double ComputeVolume(SurfaceType surface, float time) const;
 	
 private:
